fix(geocode): a shutdown stops counting as an address that cannot be found
Six of Lars's companies sat unlocated with status `failed` — every one a
valid German address that resolves at Nominatim in under a second. I
reproduced each by hand: HTTP 200, correct coordinates, no complaint.

The cause was our own shutdown. The pacer holds a lookup for up to the
policy interval before the request is even built, and it gives up when the
context does. So every lookup queued behind the current one came back with
context.Canceled the moment the worker stopped — and the worker recorded
that as a failure OF THE ADDRESS, spending one of its three attempts and
setting a day-long backoff for a request that never reached the provider.
All six failures fell inside the seven minutes I spent restarting stacks.

A cancelled lookup is now returned unrecorded. River re-queues it and the
next worker asks properly.

AND THE ROW NOW SAYS WHY. An unclassified failure publishes "the diagnosis
is in the process log", which is true and useless once the process has
restarted — exactly when somebody looks, and exactly what happened here:
the reason existed, in a log the next `make dev` had already replaced. A
provider that could not be reached is now a declared failure class with a
sentence of its own, a 503 on the wire, and the address in the log line
FaultContext writes.

The httperr mapping crossed the 500-line cap by one line, so the
ready-made refusals move to responses.go — a concept, not a line count.

Verified: reset the database, ran the backfill, killed the worker
mid-flight. Zero failed rows where the same act produced six before; 48
jobs preserved as available; the one attempt spent is a real no_match.
